modularised the repo, Docker, and start.sh
