--- excellent-articles/excellent-articles.json
@@ -1,22 +1,0 @@
-{
-"document1": "Parties agree that Confidential Information shall not be used or disclosed except as necessary for performance hereunder.",
-"document2": "Either party may terminate this Agreement by giving thirty (30) days written notice.",
-"document3": "Payment shall be made within fifteen (15) days after receipt of an invoice.",
-"document4": "Upon termination of this Agreement, each Party will return all property belonging to the other.",
-"document5": "Notices given pursuant to this Agreement must be in writing and sent via certified mail, return receipt requested.",
-"document6": "Each party agrees to indemnify and hold harmless the other from any claim arising out of their breach of this Agreement.",
-"document7": "The Parties may amend this Agreement by mutual agreement in writing.",
-"document8": "Disputes between the Parties shall be resolved through good faith negotiation.",
-"document9": "Neither party may assign its rights or obligations without prior written consent.",
-"document10": "This Agreement constitutes the entire understanding between the parties and supersedes any previous agreements.",
-"document11": "Conflicting terms in other documents will not govern unless expressly agreed to by both Parties.",
-"document12": "Any waiver of compliance with any provision of this Agreement must be in writing.",
-"document13": "Except as otherwise provided, each party bears its own costs and expenses.",
-"document14": "The Parties shall cooperate reasonably to facilitate the performance of their respective obligations.",
-"document15": "The effective date of this Agreement is the date last signed below.",
-"document16": "This Agreement shall automatically renew for successive one-year terms unless either party provides notice of non-renewal sixty (60) days before the end of the current term.",
-"document17": "In witness whereof, the parties hereto have executed this Agreement as of the Effective Date.",
-"document18": "No modification or amendment of this Agreement shall be valid or binding unless in writing and duly executed by authorized representatives of both Parties.",
-"document19": "This Agreement shall be governed by and construed in accordance with the laws of the State of New York.",
-"document20": "Any disputes arising out of or relating to this Agreement shall be referred to arbitration administered by the American Arbitration Association."
-}
