+++ gswikichat/prompt.py
@@ -1,0 +1,20 @@
+from haystack.components.builders.prompt_builder import PromptBuilder
+
+# TODO: discolm prompt https://huggingface.co/DiscoResearch/DiscoLM_German_7b_v1
+
+prompt_template = """
+Given these documents, answer the question. Answer in a full sentence. Give the response only, no explanation. Don't mention the documents.
+Documents:
+{% for doc in documents %}
+    If {{ doc.content }} answers the Question: {{question}}
+    Then return {{ doc.meta["src"] }}
+{% endfor %}
+"""
+
+# prompt_template = """
+# Given these documents, answer the question. Answer in a full sentence. Give the response only, no explanation. Don't mention the documents.
+# Documents:
+# If {{ doc.content }} answers the Question: {{question}}
+# Then only return {{ doc.meta["src"] }} and nothing at all.
+# {% endfor %}
+# """
